backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
torch==2.6.0
transformers==4.49.0
gTTS==2.5.4
deep-translator==1.11.4
Pillow==11.0.0